refactor: remove projects module
