+++ src/app/projects/project-page/project-page.component.spec.ts
@@ -26,13 +26,19 @@ describe('ProjectPageComponent', () => {
             },
           } as ProjectRouteData),
         }),
-        MockProvider(ProjectAssetsCollectionsService, {
-          byProject() {
-            return of([])
-          },
-        }),
         MockProvider(NgxMetaService),
       ],
+    })
+    TestBed.overrideComponent(ProjectPageComponent, {
+      set: {
+        providers: [
+          MockProvider(ProjectAssetsCollectionsService, {
+            byProject() {
+              return of([])
+            },
+          }),
+        ],
+      },
     })
     fixture = TestBed.createComponent(ProjectPageComponent)
     component = fixture.componentInstance
