+++ src/app/projects/projects-page/projects-page.component.spec.ts
@@ -15,13 +15,17 @@ describe('ProjectsPageComponent', () => {
         ProjectsPageComponent,
         MockComponents(ProjectListItemComponent),
       ],
-      providers: [
-        MockProvider(ProjectsService, {
-          async getListItems() {
-            return []
-          },
-        }),
-      ],
+    })
+    TestBed.overrideComponent(ProjectsPageComponent, {
+      set: {
+        providers: [
+          MockProvider(ProjectsService, {
+            async getListItems() {
+              return []
+            },
+          }),
+        ],
+      },
     })
     fixture = TestBed.createComponent(ProjectsPageComponent)
     component = fixture.componentInstance
