+++ src/app/projects/projects-page/projects-page.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core'
 import { ProjectsService } from '../projects.service'
 import { ProjectListItem } from '../project-list-item'
 import { ProjectListItemComponent } from './project-list-item/project-list-item.component'
-import { NgFor, AsyncPipe } from '@angular/common'
+import { AsyncPipe, NgFor } from '@angular/common'
 
 @Component({
   selector: 'app-projects-page',
@@ -10,6 +10,7 @@ import { NgFor, AsyncPipe } from '@angular/common'
   styleUrls: ['./projects-page.component.scss'],
   standalone: true,
   imports: [NgFor, ProjectListItemComponent, AsyncPipe],
+  providers: [ProjectsService],
 })
 export class ProjectsPageComponent {
   public readonly projects: Promise<ReadonlyArray<ProjectListItem>>
